CTCTRADERS-2836: Controllers added.
