+++ conf/app.addItems.routes
@@ -147,6 +147,9 @@ POST       /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/extra-
 GET        /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-extra-information                 controllers.addItems.previousReferences.ExtraInformationController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 POST       /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-extra-information                 controllers.addItems.previousReferences.ExtraInformationController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 
+GET        /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/check-answers              controllers.addItems.previousReferences.ReferenceCheckYourAnswersController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = NormalMode)
+GET        /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-answers             controllers.addItems.previousReferences.ReferenceCheckYourAnswersController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
+
 GET        /:lrn/add-items/:itemIndex/special-mentions/add-special-mention                       controllers.addItems.specialMentions.AddSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = NormalMode)
 POST       /:lrn/add-items/:itemIndex/special-mentions/add-special-mention                        controllers.addItems.specialMentions.AddSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = NormalMode)
 GET        /:lrn/add-items/:itemIndex/special-mentions/change-add-special-mention                controllers.addItems.specialMentions.AddSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, mode: Mode = CheckMode)
@@ -171,6 +174,9 @@ GET        /:lrn/add-items/:itemIndex/special-mentions/:referenceIndex/remove-sp
 POST       /:lrn/add-items/:itemIndex/special-mentions/:referenceIndex/remove-special-mention                        controllers.addItems.specialMentions.RemoveSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = NormalMode)
 GET        /:lrn/add-items/:itemIndex/special-mentions/:referenceIndex/change-remove-special-mention                 controllers.addItems.specialMentions.RemoveSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 POST       /:lrn/add-items/:itemIndex/special-mentions/:referenceIndex/change-remove-special-mention                 controllers.addItems.specialMentions.RemoveSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
+
+GET        /:lrn/add-items/:itemIndex/special-mentions/:referenceIndex/check-answers              controllers.addItems.specialMentions.SpecialMentionCheckYourAnswersController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = NormalMode)
+GET        /:lrn/add-items/:itemIndex/special-mentions/:referenceIndex/change-answers             controllers.addItems.specialMentions.SpecialMentionCheckYourAnswersController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 
 GET        /:lrn/add-items/:itemIndex/containers/:containerIndex/container-number                                    controllers.addItems.containers.ContainerNumberController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, containerIndex: Index, mode: Mode = NormalMode)
 POST       /:lrn/add-items/:itemIndex/containers/:containerIndex/container-number                                    controllers.addItems.containers.ContainerNumberController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, containerIndex: Index, mode: Mode = NormalMode)
@@ -227,6 +233,8 @@ POST       /:lrn/add-items/:itemIndex/documents/:documentIndex/tir-carnet-refere
 GET        /:lrn/add-items/:itemIndex/documents/:documentIndex/change-tir-carnet-reference                  controllers.addItems.documents.TIRCarnetReferenceController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, documentIndex:Index, mode: Mode = CheckMode)
 POST       /:lrn/add-items/:itemIndex/documents/:documentIndex/change-tir-carnet-reference                  controllers.addItems.documents.TIRCarnetReferenceController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, documentIndex:Index, mode: Mode = CheckMode)
 
+GET        /:lrn/add-items/:itemIndex/documents/:documentIndex/check-answers              controllers.addItems.documents.DocumentCheckYourAnswersController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, documentIndex: Index, mode: Mode = NormalMode)
+GET        /:lrn/add-items/:itemIndex/documents/:documentIndex/change-answers             controllers.addItems.documents.DocumentCheckYourAnswersController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, documentIndex: Index, mode: Mode = CheckMode)
 
 GET        /:lrn/add-items/:itemIndex/containers/:containerIndex/confirm-remove-container                        controllers.addItems.containers.ConfirmRemoveContainerController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, containerIndex: Index, mode: Mode = NormalMode)
 POST       /:lrn/add-items/:itemIndex/containers/:containerIndex/confirm-remove-container                        controllers.addItems.containers.ConfirmRemoveContainerController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, containerIndex: Index, mode: Mode = NormalMode)
