+++ conf/messages.en
@@ -800,19 +800,19 @@ addAdministrativeReference.administrativeReferenceList.label = Proof of customs 
 
 referenceType.title = What type of reference do you want to add?
 referenceType.heading = What type of reference do you want to add?
-referenceType.checkYourAnswersLabel = referenceType
+referenceType.checkYourAnswersLabel = Reference type
 referenceType.error.required = Enter the type of reference you want to add
 
 previousReference.title = What is the previous document reference?
 previousReference.heading = What is the previous document reference?
-previousReference.checkYourAnswersLabel = What is the previous document reference?
+previousReference.checkYourAnswersLabel = Previous document reference
 previousReference.error.required = Enter the previous document reference
 previousReference.error.invalidCharacters = The previous document reference must only include letters a to z without accents, numbers 0 to 9, blank spaces, at signs (@), forward slashes, back slashes, full stops, commas, hyphens, and question marks
 previousReference.error.length = The previous document reference must be 35 characters or less
 
 addExtraInformation.title = Do you need to add any extra information for this document?
 addExtraInformation.heading = Do you need to add any extra information for this document?
-addExtraInformation.checkYourAnswersLabel = addExtraInformation
+addExtraInformation.checkYourAnswersLabel = Add extra information?
 addExtraInformation.error.required = Select yes if you need to add any extra information for this document
 
 addAnotherPreviousAdministrativeReference.title.singular = You have added 1 previous administrative reference
@@ -824,7 +824,7 @@ addAnotherPreviousAdministrativeReference.error.required = Select yes if you nee
 
 extraInformation.title = What extra information for this document do you need to add?
 extraInformation.heading = What extra information for this document do you need to add?
-extraInformation.checkYourAnswersLabel = extraInformation
+extraInformation.checkYourAnswersLabel = Extra information
 extraInformation.error.required = Enter the extra information for this document
 extraInformation.error.length = The extra information for this document must be 26 characters or less
 extraInformation.error.invalid = The extra information for this document must only include letters a to z without accents, numbers 0 to 9, blank spaces, at signs (@), forward slashes, back slashes, full stops, commas, hyphens, and question marks
@@ -847,13 +847,14 @@ specialMentionType.title = What type of special mention do you want to add for i
 specialMentionType.heading = What type of special mention do you want to add for item {0}?
 specialMentionType.error.required = Enter the type of special mention that you want to add for item {0}
 specialMentionType.error.length = SpecialMentionType must be 100 characters or less
-specialMentionType.checkYourAnswersLabel = Special mention {0}
+specialMentionType.checkYourAnswersLabel = Special mention type
 
 specialMentionAdditionalInfo.title = What information do you need to add for special mention {0} for item {1}?
 specialMentionAdditionalInfo.heading = What information do you need to add for special mention {0} for item {1}?
 specialMentionAdditionalInfo.error.required = Enter the information you need to add for special mention {0} for item {1}
 specialMentionAdditionalInfo.error.length = The information you need to add for special mention {0} for item {1} must be 70 characters or less
 specialMentionAdditionalInfo.error.invalid = The information you need to add for special mention {0} for item {1} must only include letters a to z without accents, numbers 0 to 9, blank spaces, at signs (@), forward slashes, back slashes, full stops, commas, hyphens, and question marks
+specialMentionAdditionalInfo.checkYourAnswersLabel = Additional information
 
 specialMentionCheckYourAnswers.title = Check your answers
 specialMentionCheckYourAnswers.heading = Check your answers
@@ -891,6 +892,7 @@ addAnotherContainer.containerList.label = Shipping container {0}
 addExtraDocumentInformation.title = Do you need to add any extra information for document {0}?
 addExtraDocumentInformation.heading = Do you need to add any extra information for document {0}?
 addExtraDocumentInformation.error.required = Select yes if you need to add any extra information for document {0}
+addExtraDocumentInformation.checkYourAnswersLabel = Add extra information?
 
 addDocuments.title = Do you need to add any produced documents or work certificates for item {0}?
 addDocuments.heading = Do you need to add any produced documents or work certificates for item {0}?
@@ -908,7 +910,7 @@ documentReference.error.length = The document reference must be 35 characters or
 
 documentExtraInformation.title = What extra information do you need to add for document {0}?
 documentExtraInformation.heading = What extra information do you need to add for document {0}?
-documentExtraInformation.checkYourAnswersLabel = Extra information for document {0}
+documentExtraInformation.checkYourAnswersLabel = Extra information
 documentExtraInformation.error.required = Enter the extra information for document {0}
 documentExtraInformation.error.length = Document extra information must be 26 characters or less
 documentExtraInformation.error.invalid = The extra information for this document must only include letters a to z without accents, numbers 0 to 9, ampersands (&), apostrophes, at signs (@), forward slashes, full stops, hyphens, percent signs, question marks and spaces
